% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_chains_to_dataframe.R
\name{c_chains_to_dataframe}
\alias{c_chains_to_dataframe}
\title{Obtain data frame representation of measure from list of coupled chains}
\usage{
c_chains_to_dataframe(c_chains, k, m, dopar = FALSE, prune = TRUE)
}
\arguments{
\item{c_chains}{A list containing coupled chains generated by \code{\link{sample_coupled_chains}}.}

\item{k}{An integer at which to start computing the signed measure; should be less than m}

\item{m}{A time horizon: should be less than the length of the chains; typically the same
m that was used in the call to \code{\link{sample_coupled_chains}}, or a smaller value}

\item{dopar}{Boolean (default to FALSE) indicating whether to perform calculation using registered parallel cores}
}
\value{
A data frame
}
\description{
From coupled chains,
presumably generated via \code{\link{sample_coupled_chains}},
and a choice of integers k and m, the function constructs
a data frame representation of an empirical signed measure, i.e.

 \deqn{\hat{\pi}(dx) = \sum_{n=1}^N \omega_n \delta_{Z_n}(dx)}

The function returns a data frame with first column "rep" indicating index of coupled chain,
second column "MCMC" indicating whether atom is part of the "MCMC" part of the signed measure (1) or the bias correction part (0)
third column "weight" indicating weights,
remaining columns "atom.1", "atom.2", etc containing components of the atoms
}
