% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_chains_to_measure_as_list.R
\name{c_chains_to_measure_as_list}
\alias{c_chains_to_measure_as_list}
\title{Obtain empirical measure (as list) from coupled chains}
\usage{
c_chains_to_measure_as_list(c_chains, k, m)
}
\arguments{
\item{c_chains}{A list containing coupled chains generated by \code{\link{sample_coupled_chains}}.}

\item{k}{An integer at which to start computing the signed measure; should be less than m}

\item{m}{A time horizon: should be less than the length of the chains; typically the same
m that was used in the call to \code{\link{sample_coupled_chains}}, or a smaller value}
}
\value{
A list with "weights" and "atoms"
}
\description{
From coupled chains,
presumably generated via \code{\link{sample_coupled_chains}},
and a choice of integers k and m, the function constructs
a representation of an empirical signed measure, i.e.

 \deqn{\hat{\pi}(dx) = \sum_{n=1}^N \omega_n \delta_{Z_n}(dx)}

The function returns the weights \eqn{\omega} and the atoms \eqn{Z},
in a list with two entries, "weights" and "atoms".
}
