% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiasing_algorithm.R
\name{continue_coupled_chains}
\alias{continue_coupled_chains}
\title{Continue coupled MCMC chains up to K steps}
\usage{
continue_coupled_chains(c_chain, single_kernel, K = 1, ...)
}
\description{
## function to continue coupled chains until step K
c_chain should be the output of coupled_chains
and K should be more than c_chain$iteration, otherwise returns c_chain
}
