% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiasedmcmc-package.R
\docType{package}
\name{debiasedmcmc-package}
\alias{debiasedmcmc}
\alias{debiasedmcmc-package}
\title{debiasedmcmc}
\description{
...
}
\details{
...
}
\author{
Pierre E. Jacob <pierre.jacob.work@gmail.com>, John O'Leary
}
\keyword{package}

