% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgaussian_couplings.R
\name{dinvgaussian}
\alias{dinvgaussian}
\title{compute log-density of inverse Gaussian}
\usage{
dinvgaussian(x, mu, lambda)
}
\description{
at x > 0 and with given parameters mu, lambda, given by
 0.5 * log(lambda/(2*pi)) - 1.5 * log(x) - lambda * (x-mu)^2 / (2 * mu^2 * x)
}
