% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm.R
\name{fast_rmvnorm_chol}
\alias{fast_rmvnorm_chol}
\title{fast_rmvnorm_chol}
\usage{
fast_rmvnorm_chol(nparticles, mean, chol)
}
\arguments{
\item{mean}{A vector of size d specifying the mean vector of the multivariate Normal}

\item{chol}{A matrix of size d x d specifying the upper triangular Cholesky factor
of the covariance matrix of the multivariate Normal target,
for instance obtained using the \code{\link[base]{chol}}
function of R.}

\item{n}{An integer >= 1 specifying the desired number of draws}
}
\value{
A matrix of size n x d containing n d-dimensional multivariate Normal draws (one per row)
}
\description{
Generate multivariate Normal draws. The function does not check
the arguments, use at your own risk.
}
\examples{
Sigma <- diag(1, 5, 5)
Sigma[1,2] <- Sigma[2,1] <- 0.3
fast_rmvnorm_chol(2, rep(0, 5), chol(Sigma))
}
