% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_couplings.R
\name{gaussian_max_coupling}
\alias{gaussian_max_coupling}
\title{Maximal coupling of two multivariate Normal distributions}
\usage{
gaussian_max_coupling(mu1, mu2, Sigma1, Sigma2)
}
\description{
sample from maximal coupling of two multivariate Normal distributions,
specified through their means and covariance matrices
}
