% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_couplings.R
\name{gaussian_opt_transport}
\alias{gaussian_opt_transport}
\title{Optimal transport coupling between two multivariate Normals}
\usage{
gaussian_opt_transport(nsamples, m1, m2, S1_chol, S2_chol, S1_chol_inv,
  S2_chol_inv)
}
\description{
sample from optimal transport coupling of two multivariate Normal distributions
}
