% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_coupling.R
\name{get_max_coupling}
\alias{get_max_coupling}
\title{Sample from maximally coupled distributions p and q}
\usage{
get_max_coupling(rp, dp, rq, dq)
}
\description{
Takes two distributions (sampling function and pdf),
and return a function to sample from a maximal coupling of these distributions
}
