% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mh_kernels.R
\name{get_mh_kernels}
\alias{get_mh_kernels}
\title{Get random walk Metropolis-Hastings kernels}
\usage{
get_mh_kernels(logtarget, Sigma_proposal, dimension)
}
\arguments{
\item{logtarget}{function to compute target log-density}

\item{Sigma_proposal}{variance of the Normal random walk for the proposals}

\item{dimension}{dimension of the target distribution}
}
\value{
A list containing the keys
\code{single_kernel}, \code{coupled_kernel}.
}
\description{
This function takes descriptions of the target
and tuning parameters, and returns a list containing the keys
\code{single_kernel}, \code{coupled_kernel} corresponding to marginal
and coupled MH kernels, with Normal random walks.
These kernels can then be used in the function \code{\link{coupled_chains}}.
}
