% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableselection.R
\name{get_variableselection}
\alias{get_variableselection}
\title{Y and X need to be matrices, and lambda non-negative}
\usage{
get_variableselection(Y, X, g, kappa, s0, proportion_singleflip)
}
\description{
Y and X need to be matrices, and lambda non-negative
}
