% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_bar.R
\name{H_bar}
\alias{H_bar}
\title{Compute unbiased estimators from coupled chains}
\usage{
H_bar(c_chains, h = function(x) x, k = 0, m = 1)
}
\arguments{
\item{c_chains}{A list containing coupled chains generated by \code{\link{sample_coupled_chains}}.}

\item{h}{A test function of interest, which should take a chain state ("chain_state" entry of the output of "rinit", for instance)
and return a numeric vector}

\item{k}{An integer at which to start computing the unbiased estimator; should be less than m}

\item{m}{A time horizon: should be less than the length of the chains; typically the same
m that was used in the call to \code{\link{sample_coupled_chains}}, or a smaller value}
}
\value{
A value (or vector of values) of an unbiased estimator of \eqn{\int h(x) \pi(x) dx}
}
\description{
Compute unbiased estimators based on coupled chains.
Presumably generated via \code{\link{sample_coupled_chains}}.

The test function h should take a "chain_state" as argument and return a numeric vector.
The estimand of interest is \eqn{\int h(x) \pi(x) dx}, where \eqn{\pi} is the invariant distribution
of the chains.

The lag is inferred from the coupled chains, so there's no argument to specify it.
}
