% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_c_chains.R
\name{histogram_c_chains}
\alias{histogram_c_chains}
\title{histogram_c_chains}
\usage{
histogram_c_chains(c_chains, component, k, m, breaks = NULL,
  nclass = 30)
}
\description{
Compute histograms based on coupled Markov chains
}
