% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{logisticregression_precomputation}
\alias{logisticregression_precomputation}
\title{Precomputation to prepare for the Polya-Gamma sampler}
\usage{
logisticregression_precomputation(Y, X, b, B)
}
\description{
This function takes the canonical elements defining the logistic regression
problem (the vector of outcome Y, covariate matrix X, the prior mean b and the prior variance B),
and precomputes some quantities repeatedly used in the Polya-Gamma sampler and variants of it.
The precomputed quantities are returned in a list, which is then meant to be passed to the samplers.
}
