% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{pg_gibbs}
\alias{pg_gibbs}
\title{Polya-Gamma Gibbs sampler}
\usage{
pg_gibbs(niterations, logistic_setting)
}
\value{
a matrix where each row corresponds to an iteration of the sampler, and contains
the regression coefficient at that iteration.
}
\description{
This implements the sampler proposed in
Nicholas G Polson, James G Scott, and Jesse Windle. Bayesian inference for logistic models using
Polya–Gamma latent variables. Journal of the American statistical Association, 108(504):1339–1349, 2013.
The arguments are:
 \itemize{
  \item niterations: the number of desired MCMC iterations,
  \item logistic_setting: a list of precomputed quantities obtained via 'logistic_precomputation'.
  }
}
