% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_couplings.R
\name{rgamma_coupled}
\alias{rgamma_coupled}
\title{Sample from maximally coupled Gamma}
\usage{
rgamma_coupled(alpha1, alpha2, beta1, beta2)
}
\description{
with given parameters alpha1, alpha2, beta1, beta2,
where the parametrization is that beta is the rate, i.e.
alpha * log(beta) - lgamma(alpha) + (alpha-1) * log(x) - beta x
}
