% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgamma_couplings.R
\name{rigamma_coupled}
\alias{rigamma_coupled}
\title{Sample from maximally coupled inverse gamma}
\usage{
rigamma_coupled(alpha1, alpha2, beta1, beta2)
}
\description{
with given parameters alpha1, alpha2, beta1, beta2,
where the parametrization is that the log-density of IG(alpha, beta) is
alpha * log(beta) - lgamma(alpha) - (alpha+1) * log(x) - beta / x
}
