% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgamma_couplings.R
\name{rinversegamma}
\alias{rinversegamma}
\title{Sample from inverse gamma}
\usage{
rinversegamma(n, alpha, beta)
}
\description{
with given parameters alpha, beta, with log-density given by
alpha * log(beta) - lgamma(alpha) - (alpha+1) * log(x) - beta / x
}
