% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgaussian_couplings.R
\name{rinvgaussian}
\alias{rinvgaussian}
\title{Sample from inverse Gaussian}
\usage{
rinvgaussian(n, mu, lambda)
}
\value{
A vector of n draws, where n is the first argument.
}
\description{
Parametrized by mu, lambda, with log-density given by
 \deqn{0.5 * log(\lambda/(2*\pi)) - 1.5 * log(x) - \lambda * (x-\mu)^2 / (2 * \mu^2 * x)}

 The procedure goes as follows.

 \itemize{
 \item Generate nu ~ Normal(0,1).
 \item Define y = nu^2.
 \item Define x = mu + mu^2 * y / (2 * lambda) - mu / (2 * lambda) * sqrt(4 * mu * lambda * y + mu^2 * y^2).
 \item Generate Z ~ Uniform(0,1).
 \item If z <= mu / (mu + x), output x, otherwise output mu^2 / x.
 }
}
