% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_couplings.R
\name{rmvnorm_max}
\alias{rmvnorm_max}
\title{Maximal coupling of two multivariate Normal distributions}
\usage{
rmvnorm_max(mu1, mu2, Sigma1, Sigma2)
}
\description{
sample from maximal coupling of two multivariate Normal distributions,
specified through their means and covariance matrices
}
