% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_couplings.R
\name{rmvnorm_max_chol}
\alias{rmvnorm_max_chol}
\title{Maximal coupling of two multivariate Normal distributions}
\usage{
rmvnorm_max_chol(mu1, mu2, Cholesky1, Cholesky2, Cholesky_inverse1,
  Cholesky_inverse2)
}
\description{
sample from maximal coupling of two multivariate Normal distributions,
specified through their means, the Cholesky factors of their covariance matrices,
and the Cholesky factors of the inverse covariance matrices (i.e. the precision matrices)
}
