% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_couplings.R
\name{rmvnorm_reflectionmax}
\alias{rmvnorm_reflectionmax}
\title{Reflection-Maximal coupling of two multivariate Normal distributions}
\usage{
rmvnorm_reflectionmax(mu1, mu2, Cholesky, Cholesky_inverse)
}
\description{
sample from reflection-maximal coupling of two multivariate Normal distributions,
specified through their means, with the same covariance matrix specified
through its Cholesky factor and Cholesky factor of its inverse.
}
