% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_couplings.R
\name{rnorm_max_coupling}
\alias{rnorm_max_coupling}
\title{Maximal coupling of two univariate Normal distributions}
\usage{
rnorm_max_coupling(mu1, mu2, sigma1, sigma2)
}
\description{
sample from maximal coupling of two univariate Normal distributions,
specified through their means and standard deviations
}
