% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled_chains.R
\name{sample_coupled_chains}
\alias{sample_coupled_chains}
\title{Sample coupled Markov chains}
\usage{
sample_coupled_chains(single_kernel, coupled_kernel, rinit, m = 1,
  lag = 1, max_iterations = Inf, preallocate = 10)
}
\description{
sample two Markov chains, each following 'single_kernel' marginally,
and 'coupled_kernel' jointly, until min(max(tau, m), max_iterations), where tau
is the first time the two chains meet. Or more precisely, they meet with a delay of one, i.e. X_t = Y_{t-1}.
}
