% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetingtime.R
\name{sample_meetingtime}
\alias{sample_meetingtime}
\title{Sample coupled Markov chains until meeting}
\usage{
sample_meetingtime(single_kernel, coupled_kernel, rinit, lag = 1,
  max_iterations = Inf)
}
\arguments{
\item{single_kernel}{A list taking a state and returning a state, performing one step of a Markov kernel}

\item{coupled_kernel}{A list taking two states and returning two states, performing one step of a coupled Markov kernel;
it also returns a boolean "identical" indicating whether the two states are identical.}

\item{rinit}{A list representing the initial state of the chain, that can be given to 'single_kernel'}

\item{lag}{A time lag, equal to one by default}

\item{max_iterations}{A maximum number of iterations, at which to interrup the while loop; Inf by default}
}
\value{
A list with
\itemize{

\item meetingtime: the meeting time; equal to Inf if while loop was interrupted

\item elapsedtime: elapsed wall-clock time, in seconds
}
}
\description{
Sample two Markov chains, each following 'single_kernel' marginally,
until they meet, and report the meeting time, as well as the elapsed wall-clock time in seconds.

This function does not record the trajectories of the chains, with the goal of being memory-light.
To record these trajectories, see \code{\link{sample_coupled_chains}}. To directly
compute unbiased estimators on the fly, see  \code{\link{sample_unbiasedestimator}}.
}
