% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_coupling.R
\name{get_max_coupling}
\alias{get_max_coupling}
\title{Sample from maximal coupling of two distributions p and q}
\usage{
get_max_coupling(rp, dp, rq, dq)
}
\arguments{
\item{rp}{A function taking n as an argument and returning n samples from the distribution p}

\item{dp}{A function taking x as an argument and returning log-pdf of p evaluated at x}

\item{rq}{A function taking n as an argument and returning n samples from the distribution q}

\item{dq}{A function taking x as an argument and returning log-pdf of q evaluated at x}
}
\value{
Returns a list with

\itemize{
\item "xy": the pair of samples \eqn{(x,y)}

\item "identical": TRUE if \eqn{x = y}, FALSE otherwise
}
}
\description{
Takes two univariate continuous distributions (specified by random number generator and log-pdf function),
and returns a function to sample from a maximal coupling of these two distributions.
}
\examples{
mu1 <- 0; mu2 <- 1; sigma1 <- 0.5; sigma2 <- 1.2
 f <- get_max_coupling(function(n) rnorm(n, mu1, sigma1),
 function(x) dnorm(x, mu1, sigma1, log = TRUE),
 function(n) rnorm(n, mu2, sigma2),
 function(x) dnorm(x, mu2, sigma2, log = TRUE))
f()
}
