% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mh_kernels.R
\name{get_mh_kernels}
\alias{get_mh_kernels}
\title{Get random walk Metropolis-Hastings kernels}
\usage{
get_mh_kernels(target, Sigma_proposal)
}
\arguments{
\item{target}{function taking a vector as input and returning target log-density evaluation}

\item{Sigma_proposal}{covariance of the Normal random walk proposal}
}
\value{
A list containing the keys
\code{single_kernel}, \code{coupled_kernel}.
}
\description{
This function takes a target (specified through its log-pdf)
and a covariance matrix for a Normal random walk proposal, and returns a list containing the keys
\code{single_kernel}, \code{coupled_kernel} corresponding to marginal
and coupled MH kernels.
These kernels can then be used in the functions \code{\link{sample_meetingtime}} or
\code{\link{sample_coupled_chains}} or \code{\link{sample_unbiasedestimator}}
}
