% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiasing_algorithm.R
\name{H_bar}
\alias{H_bar}
\title{Compute unbiased estimators from coupled chains}
\usage{
H_bar(c_chains, h = function(x) x, k = 0, K = 1)
}
\description{
Compute the proposed unbiased estimators, for each of the element
in the list 'c_chains'. The integral of interest is that of the function h,
which can be multivariate. The estimator uses the variance reduction technique
whereby the estimator is the MCMC average between times k and K, with probability
going to one as k increases.
}
