% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgaussian_couplings.R
\name{rinvgaussian_coupled}
\alias{rinvgaussian_coupled}
\title{Sample from maximally coupled inverse Gaussian}
\usage{
rinvgaussian_coupled(mu1, mu2, lambda1, lambda2)
}
\value{
A pair of values in a vector of size two.
}
\description{
with parameters mu1, mu2, lambda1, lambda2; see \code{\link{rinvgaussian}}.
}
