% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_max_coupling.R
\name{rnorm_max_coupling}
\alias{rnorm_max_coupling}
\title{Maximal coupling of two univariate Normal distributions}
\usage{
rnorm_max_coupling(mu1, mu2, sigma1, sigma2)
}
\arguments{
\item{mu1}{First mean}

\item{mu2}{Second mean}

\item{sigma1}{First mean}

\item{sigma2}{Second mean}
}
\value{
Returns a list with

\itemize{
\item "xy": the pair of samples \eqn{(x,y)}

\item "identical": TRUE if \eqn{x = y}, FALSE otherwise
}
}
\description{
Sample from maximal coupling of two univariate Normal distributions,
specified through their means and standard deviations. See \code{\link{rmvnorm_max}} for a multivariate version.
}
